Correciones al reto 4

+++ Reto4/src/main/java/utp/misiontic2022/c2/p47/reto4/modelo/dao/LiderDao.java
@@ -33,7 +33,7 @@ public class LiderDao {
             lider.setNombre(resultados.getString("Nombre"));
             lider.setPrimer_Apellido(resultados.getString("Primer_Apellido"));
             lider.setSegundo_Apellido(resultados.getString("Segundo_Apellido"));
-            lider.setSalario(resultados.getInt("Salario"));
+            lider.setSalario(resultados.getFloat("Salario"));
             lista.add(lider);
         }
         return lista;
